backend/project/requirements.txt
# ─── Core ────────────────────────────────────────────────────────────────────
Django>=5.0,<6.1
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0

# ─── Database ────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9,<3.0

# ─── Media / Images ─────────────────────────────────────────────────────────
Pillow>=10.0,<11.0

# ─── Production Server ──────────────────────────────────────────────────────
gunicorn>=21.2,<23.0

# ─── Environment Variables ──────────────────────────────────────────────────
python-dotenv>=1.0,<2.0
